+++ assignment/week_3/RecommendedMenu/src/components/OnBoarding.jsx
@@ -1,13 +1,10 @@
 import { useState } from "react";
 import styled from "styled-components";
 import StartRecommend from "./StartRecommend";
+import SelectCategory from "./SelectCategory";
 
 const OnBoarding = () => {
   const [selectedCategory, setSelectedCategory] = useState("");
-
-  const clickCategoryHandler = (e) => {
-    setSelectedCategory(e.target.innerHTML);
-  };
 
   return (
     <St.Container>
@@ -18,22 +15,7 @@ const OnBoarding = () => {
       {selectedCategory ? (
         <StartRecommend selectedCategory={selectedCategory} />
       ) : (
-        <St.CategoryContainer>
-          <St.Category
-            onClick={(e) => {
-              clickCategoryHandler(e);
-            }}
-          >
-            취향대로 추천
-          </St.Category>
-          <St.Category
-            onClick={(e) => {
-              clickCategoryHandler(e);
-            }}
-          >
-            랜덤 추천
-          </St.Category>
-        </St.CategoryContainer>
+        <SelectCategory setSelectedCategory={setSelectedCategory} />
       )}
     </St.Container>
   );
@@ -61,35 +43,6 @@ const St = {
     border-radius: 1rem;
 
     background-color: ${({ theme }) => theme.colors.lightYellow};
-  `,
-  CategoryContainer: styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    width: 55rem;
-    padding: 1rem 0;
-
-    gap: 1rem;
-  `,
-  Category: styled.button`
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    width: 50%;
-    height: 20rem;
-    border-radius: 3rem;
-
-    word-break: keep-all;
-
-    font-size: 1.7rem;
-    background-color: ${({ theme }) => theme.colors.lightYellow};
-
-    &:hover {
-      color: ${({ theme }) => theme.colors.white};
-      background-color: ${({ theme }) => theme.colors.darkGreen};
-    }
   `,
 };
